backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
python-dotenv==1.2.2
Pillow==12.1.1
django-filter==24.1
django-cors-headers==4.3.1
PyJWT>=2.8.0,<3.0.0
firebase-admin==6.3.0
gunicorn==21.2.0
